+++ logs/formatLogLine.js
