+++ logs/printLogEvent.js
